Add a Memory Object implementation
The memory object is a struct containing a pointer and a pointer to
a function used for de-allocation.

Ownership can be transferred on setting the de-allocator, in which case
the de-allocator function will be called to free the memory object.
Providing a NULL de-allocator leaves the pointer alone, resource handling
must then be done somewhere else.