doc: Introduce Doxygen
This adds a new CMake custom target if Doxygen is found on the build
host. The Doxyfile will be generated in the build folder from the new
Doxyfile.in which contains placeholders with CMake variables which are
replaced by CMake.

Doxygen style file headers were added to the library header files to
actually create a first more or less useful Doxygen HTML output.

Note: Doxygen still creates a lot of warnings about not yet documented
functions and structs.