travis: Simplify config
Travis did some updates since 2017, when we last touched our .travis.yml
and now things started to break due to packages not available somehow.

And: all that matrix configuration is complicated and we don't have to
compile it with every compiler of the world, so we switch back to a
simpler configuration for now.

This basically reverts 2add3afe437b ("ci: travis: Replace compiler by
matrix"), but also removes explicit pin down to Ubuntu 14.04 (trusty).