build: lib: Use target dependent include directories
In so called Modern CMake it is good practice to avoid global
manipulation of include directories and set include directories as
target properties. This way those can be exported later and user can
just depend on the namespaced target names. We declare those as alias
for internal use (we will need that e.g. for unit tests).